attempt to fix #295 on old safari mac machines

+++ src/glsl/webgl2/catalogs/healpix.vert
@@ -1,5 +1,6 @@
 #version 300 es
-precision lowp float;
+precision highp float;
+
 layout (location = 0) in vec2 offset;
 layout (location = 1) in vec2 uv;
 layout (location = 2) in vec3 center;
