+++ src/glsl/webgl2/hips/raytracer/i16.frag
@@ -1,7 +1,7 @@
 #version 300 es
-precision lowp float;
-precision lowp sampler2DArray;
-precision mediump int;
+precision highp float;
+precision highp sampler2DArray;
+precision highp int;
 
 uniform sampler2DArray tex;
