+++ src/glsl/webgl2/hips/raytracer/u8.frag
@@ -1,9 +1,9 @@
 #version 300 es
-precision lowp float;
-precision lowp sampler2DArray;
-precision lowp usampler2DArray;
-precision lowp isampler2DArray;
-precision mediump int;
+precision highp float;
+precision highp sampler2DArray;
+precision highp usampler2DArray;
+precision highp isampler2DArray;
+precision highp int;
 
 uniform sampler2DArray tex;
